Move migration code away from activity into separate job.
